split serial tests
